Add logical minimum and maximum to HID descriptor for analog axes.
Reporting a logical minimum (0) and maximum (2**12 - 1 = 4095)
that matches the range of the 12-bit A2D allows the host to scale
the analog inputs automatically, allowing use of the analog axes
without calibration.

This change required adding the LogicalMaximum16b macro, which
encodes a 2-byte logical maximum for use in the descriptor. (The
existing LogicalMaximum macro encodes a 1-byte value.)